requirements.txt
aiogram==3.4.1
sqlalchemy==2.0.28
aiosqlite==0.20.0
apscheduler==3.10.4
python-dotenv==1.0.1
pdfplumber==0.10.0
openpyxl==3.1.2
gspread==6.0.0
google-auth==2.27.0
